# api/openapi.yaml
openapi: 3.1.0
info:
  title: one-box FaaS REST API
  version: 1.0.0
  summary: Customer-facing REST API for the one-box FaaS platform.
  description: |
    Public REST surface exposed by `apid` on `127.0.0.1:8081` (proxied by
    `gatewayd` at the public listener). Every route is authenticated with a
    Bearer API key (`fp_live_…`) and rate-limited per-IP (10 failed-auth / min).
    Successful POSTs accept an `Idempotency-Key` header (24h TTL) and replay
    with `Idempotent-Replayed: true`.

    Errors use RFC 7807 `application/problem+json` with a stable `code`
    field — see `Problem` schema. The single documented exception is the
    429 response from the authlimiter middleware: plain text only
    (`pkg/middleware/authlimit.go`).

    Operator routes (`/v1/compute-nodes`), dashboard routes (`/dashboard/*`),
    webhook receivers (`/v1/webhooks/stripe`), SSE streams (`/v1/events`,
    `/v1/deployments/{id}/logs`), and the CLI device-code flow
    (`/v1/cli-auth/*`) are not part of this public spec.
  contact:
    name: one-box FaaS
    url: https://docs.gregale.dev
  license:
    name: Proprietary
    identifier: LicenseRef-onebox-internal
servers:
  - url: https://{host}
    description: Public endpoint (proxied through gatewayd).
    variables:
      host:
        default: api.gregale.dev
        description: Public hostname (one-box FaaS).
...
security:
  - bearerAuth: []
paths:
...
  /login:
    post:
      tags: [auth]
      operationId: passwordLogin
      summary: Email + password sign-in.
      description: |
        Verifies the email + password against the `account_passwords`
        row (Argon2id, PHC string). Sets the `faas_sid` session
        cookie on success. The response body carries only
        `{account_id, plan}` — no API key. The SDK's device-code
        flow is the programmatic-auth path; the dashboard cookie is
        the only auth artifact on the browser side.

        Anti-enumeration: unbound email, wrong password, and
        passwordless (OAuth-only) accounts all return 401
        `invalid_credentials` with the same body. Every code path
        runs one Argon2id verify under identical parameters so the
        timing oracle stays closed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordLoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordLoginRequest'
      responses:
        '200':
          description: |
            Signed in. The `Set-Cookie: faas_sid=…` header carries
            the session cookie.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: |
                `faas_sid=<sealed-session>; HttpOnly; Secure;
                SameSite=Lax`. 7-day fixed lifetime.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordLoginResponse'
                example: {"account_id":"0123456789abcdef0123456789abcdef","plan":"hobby"}
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /signup:
    post:
      tags: [auth]
      operationId: passwordSignup
      summary: Create account with email + password.
      description: |
        Creates an account and signs the caller in. On a colliding
        email + matching password, the call is idempotent and signs
        in. On a colliding email + different password, the response
        is 401 `invalid_credentials` (never 409 — the surface
        forbids account enumeration).

        Password floor: 12 characters (NIST-style; no complexity
        rules).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordSignupRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordSignupRequest'
      responses:
        '200':
          description: Account created (or reused) and signed in.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: |
                `faas_sid=<sealed-session>; HttpOnly; SameSite=Lax`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordLoginResponse'
        '400':
          description: Password too weak (≥12 chars required).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /login/forgot:
    post:
      tags: [auth]
      operationId: passwordForgot
      summary: Request a password-reset email.
      description: |
        Always returns 200 with the same body regardless of
        whether the email is bound to an account. The reset URL
        is mailed via the platform's outbound mailer; the response
        never leaks account presence.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Request accepted. The reset email is sent if the address is registered.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [ok] }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/reset:
    get:
      tags: [auth]
      operationId: passwordResetForm
      summary: Render the password-reset form.
      description: |
        Validates the token shape; renders the
        `password_reset_form.html` template on a valid token.
        Invalid / expired / consumed tokens return 410 Gone.
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string, pattern: '^[A-Za-z0-9_-]{43}$' }
          description: |
            Base64url-encoded 32-byte token from the reset email.
      responses:
        '200':
          description: Token is valid; renders the form.
          content:
            text/html:
              schema: { type: string }
        '410': { $ref: '#/components/responses/ResetTokenInvalid' }
    post:
      tags: [auth]
      operationId: passwordResetConfirm
      summary: Submit a new password against a reset token.
      description: |
        Atomically consumes the token (one-shot, replay returns
        410), Argon2id-encodes the new password, sets it on the
        account, and signs the caller in.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
      responses:
        '302':
          description: |
            Password set. Redirects to `/dashboard/`. `Set-Cookie:
            faas_sid=…` is set.
          headers:
            Location:
              schema: { type: string, example: "/dashboard/" }
        '400':
          description: New password is too weak (≥12 chars required).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '410': { $ref: '#/components/responses/ResetTokenInvalid' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /v1/auth/logout:
    post:
      tags: [auth]
      operationId: postAccountLogout
      summary: Log the current dashboard session out.
      description: |
        Revokes the calling session row in the `sessions` table
        (one row per dashboard login; ADR-039 / IAM-3) and
        clears the `faas_sid` cookie. Always 204 on success,
        even if the row was already revoked (idempotent — the
        cookie is cleared either way). Other sessions on the
        same account are NOT touched; use
        `POST /v1/auth/sessions/revoke_all` for that.

        CSRF: action `logout` (verify via the
        `faas_csrf` cookie + body `csrf_token` field or
        `X-CSRF-Token` header).

        Emits `auth.session.revoke` with
        `reason: "logout"`.
      parameters:
        - $ref: '#/components/parameters/CookieSession'
      responses:
        '204':
          description: |
            Logout successful. `Set-Cookie: faas_sid=…` clears
            the session cookie.
        '400':
          description: CSRF check failed.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401':
          description: |
            Cookie missing, expired, tampered, or its `sid` is
            not backed by an active sessions row. The cookie
            is cleared on the wire. Surfaces as
            `session_expired` (pre-IAM-3 cookie / missing /
            revoked) or `session_invalid` (account-mismatch
            defensive path; AEAD-bound envelopes should not
            produce this).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /dashboard/account/set-password:
    post:
      tags: [auth]
      operationId: setPassword
      summary: Set a password on the authenticated account.
      description: |
        Authenticated. Lets OAuth-only customers opt into password
        login. The same Argon2id path as `/auth/reset`, anchored
        to the session's account rather than a reset token.
      parameters:
        - $ref: '#/components/parameters/CookieSession'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetPasswordRequest'
      responses:
        '302':
          description: Password set. Redirects to `/dashboard/account/`.
          headers:
            Location:
              schema: { type: string, example: "/dashboard/account/" }
        '400':
          description: Chosen password is too weak (≥12 chars required).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ─── Apps ─────────────────────────────────────────────────────────────────
...
  /v1/deployments/{id}/logs:
    parameters:
      - $ref: '#/components/parameters/Id32'
    get:
      tags: [deployments]
      operationId: streamDeploymentLogs
      summary: Stream build logs (SSE).
      description: |
        Server-Sent Events stream of build logs. `follow=1` holds the
        connection open until the build completes.
      parameters:
        - name: before_seq
          in: query
          schema: { type: integer }
          description: Stream cursor — return only log entries with seq strictly less than this value.
        - name: limit
          in: query
          schema: { type: integer }
          description: Maximum number of log entries to return in the initial burst before streaming.
        - name: follow
          in: query
          schema: { type: integer, enum: [0, 1], default: 0 }
          description: If 1, hold the connection open and stream new build log entries as they arrive.
      responses:
        '200':
          description: A text/event-stream of build log entries, terminated by an empty SSE frame when the build finishes.
          content:
            text/event-stream: {}
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  # ─── Builds (ADR-038, Tier 3 / issue #197 B3.10-read half) ───────────────
  # The provenance route is the only /v1/builds surface today.
  # deployments.id remains the parent resource; Build:read scope
  # (api.ScopesReadSurface) gates the read.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer API key. Format: `fp_live_` + 48 hex chars. The plaintext is
        returned exactly once on `POST /v1/keys`; subsequent reads only
        surface the prefix.
...
  schemas:
...
    PasswordLoginRequest:
      description: "Email + password for sign-in. Email is canonicalised server-side (lowercase + trim)."
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 256 }
...